[#48] refactor: banner 데이터 조회 product 도메인 -> banner 도메인 내에서 구현
