+++ src/main/java/com/napzak/domain/product/api/exception/ProductSuccessCode.java
@@ -10,9 +10,7 @@ import org.springframework.http.HttpStatus;
 public enum ProductSuccessCode implements BaseSuccessCode {
     RECOMMEND_PRODUCT_GET_SUCCESS(HttpStatus.OK, "개인화 상품 불러오기 성공"),
     TOP_SELL_PRODUCT_GET_SUCCESS(HttpStatus.OK, "팔아요 인기 상품 불러오기 성공"),
-    TOP_BUY_PRODUCT_GET_SUCCESS(HttpStatus.OK, "구해요 인기 상품 불러오기 성공"),
-    BANNER_GET_SUCCESS(HttpStatus.OK, "배너 정보 불러오기 성공")
-    ;
+    TOP_BUY_PRODUCT_GET_SUCCESS(HttpStatus.OK, "구해요 인기 상품 불러오기 성공");
 
     private final HttpStatus httpStatus;
     private final String message;
